Questao 3

+++ src/lista_avaliativa_1/questao4.c
@@ -3,7 +3,7 @@
 int main()
 {
   char tipo;
-  float salario;
+  double salario;
 
   scanf("%c", &tipo);
   scanf("%f", &salario);
